Html5 game support

+++ MakerV3.py
@@ -1,10 +1,11 @@
-import os , platform , string , time , getpass 
+import os , platform , string , time , getpass , stat
 
 #----------------------------------------------------------
 #variables
 #For all of the "_"'s in the script that is an unused variable and pylint doesn't care so that's what I used
 UserName = getpass.getuser()
 SkipPath = '         '
+emptyport = 8000
 OS = platform.system()# for whatever reason when I ported this to macOS catalina platform.system returned Darwin so macOS = Darwin in python
 
 Blacklist = open("info/blacklist.txt" , 'r')
@@ -17,21 +18,20 @@ def split_path(path):
     Path = path
     if OS == "Windows":
         if path.endswith("index.html"):
-            name , _ = Path.split(".")
-            S , name , _ = name.rsplit("\\",2)
-            start = os.path.join(S , name)
+            return (AddHTMLGame(path))
         else:
             start, name = path.rsplit("\\", 1)
             name , _ = (name.split('.', 1))
     if OS == "Linux":
         start, name = path.rsplit("/", 1)
-        if path.endswith("index.html"):
-            _ , name , _ = path.rsplit("/",2)
-            start , _= path.rsplit("/", 1)
+        if Path.endswith("index.html"):
+            return(AddHTMLGame(path))
         if "." in name:
             name , _ = (name.split('.', 1))
     if OS == "Darwin":
         global SkipPath
+        if Path.endswith('index.html'):
+            return(AddHTMLGame(path))
         ActualPath , _ = path.split(".app",1)
         Path = ActualPath+".app"
         if SkipPath == '         ':
@@ -145,7 +145,6 @@ def GetInstallLocation():
                                 continue
                             if ("userdata" in CorrectDir):
                                 foundit = 0
-                                print("found it")
                                 SteamLocal = Root
                                 print ('\nfound steam install in "%s"'%SteamLocal)
                                 continue
@@ -160,12 +159,12 @@ def GetInstallLocation():
                     continue
     return SteamIDnum , SteamLocal
 def getsettings():
-    global SteamID , InstallLocation , DefaultCleanout , Proton
+    global SteamID , InstallLocation , DefaultCleanout , Proton , EnableHTML
     #settings are layed out like "SteamID , Steam Install Location , cleanout by default , Enable Proton(for running windows games on linux through steam)"
     SettingFile = open("info/settings" , 'r')
     SavedSettings = SettingFile.read()
-    SteamID , InstallLocation , DefaultCleanout , Proton = SavedSettings.split(' , ')
-    if SavedSettings == "'' , '' , '' , ''":
+    SteamID , InstallLocation , DefaultCleanout , Proton , EnableHTML = SavedSettings.split(' , ')
+    if SavedSettings == "'' , '' , '' , '' , ''":
         print ("\n\nLooks like you have never used this tool before (or you downloaded a new version or something)\nlets go through some setup.\n")
         time.sleep(0.5)
         properanswer = 0
@@ -191,11 +190,20 @@ def getsettings():
                 if EnableProton == ("n"):
                     Proton = "no"
                     properanswer2 = 1
+        properanswer = 0
+        while properanswer == 0:
+            EnableHTML = input("would you like to enable HTML5 games?(this is more or less done but has not been extensively bug tested) (y/n)")
+            if EnableHTML == ("y"):
+                EnableHTML = "yes"
+                properanswer = 1
+            if EnableHTML == ("n"):
+                EnableHTML = "no"
+                properanswer = 1
         GetInstallLocation()
         SettingsWrite = open("info/settings" , 'w')
         SteamID = SteamID.replace(SteamID , SteamIDnum)
         InstallLocation = InstallLocation.replace(InstallLocation , SteamLocal)
-        FullSettings = ('%s , "%s" , %s , %s'%(SteamID , InstallLocation , Cleanout , Proton))
+        FullSettings = ('%s , "%s" , %s , %s , %s'%(SteamID , InstallLocation , Cleanout , Proton , EnableHTML))
         SettingsWrite.write(FullSettings)
     if DefaultCleanout == 'y':
         DefaultCleanout = Cleanout
@@ -203,7 +211,7 @@ def getsettings():
         InstallLocation = SteamLocal
     if open('info/Dirs.txt','r').read() == '':
         NoDirs()
-    return SteamID , InstallLocation , DefaultCleanout , Proton
+    return SteamID , InstallLocation , DefaultCleanout , Proton , EnableHTML
 def Cleanout():
     global ReplaceVDF
     if OS == "Windows":
@@ -229,7 +237,8 @@ def Cleanout():
                     if file == 'shortcuts.vdf':
                         ReplaceVDF = os.path.join(Root , file)
         if DefaultCleanout == 'yes':
-            os.popen('rm "%s/shortcuts.vdf"'%(ReplaceVDF))
+            if os.path.exists("%s/shortcuts.vdf"%ReplaceVDF):
+                os.popen('rm "%s/shortcuts.vdf"'%(ReplaceVDF))
             os.system('cp "%s" "%s"'%(BaseVDF , ReplaceVDF))
     if OS == "Darwin":
         BaseVDF = "info/shortcuts.vdf"
@@ -341,13 +350,14 @@ def getfiles(scandir):
                     ExeFile = (os.path.join(Root , file))
                     InBlacklist(ExeFile)
                 if file == "index.html":
-                    if OldRoot in os.path.join(Root , file):
-                        pass
-                    else:
-                        _ , GameName , _ = Root.rsplit("\\",2)
-                        OldRoot = GameName
-                        HTMLFILE = (os.path.join(Root , file))
-                        InBlacklist(HTMLFILE)
+                    if EnableHTML == "yes":
+                        if OldRoot in os.path.join(Root , file):
+                            pass
+                        else:
+                            _ , GameName , _ = Root.rsplit("\\",2)
+                            OldRoot = GameName
+                            HTMLFILE = (os.path.join(Root , file))
+                            InBlacklist(HTMLFILE)
             if OS == "Linux":
                 CheckFile = os.path.join(Root , file)
                 ExeCheck = os.access(CheckFile, os.X_OK)
@@ -355,13 +365,14 @@ def getfiles(scandir):
                     ExecFile = (os.path.join(Root , file))
                     InBlacklist(ExecFile)
                 if file == "index.html":
-                    if OldRoot in os.path.join(Root , file):
-                        pass
-                    else:
-                        _ , GameName , _ = Root.rsplit("/",2)
-                        OldRoot = GameName
-                        HTMLFILE = (os.path.join(Root , file))
-                        InBlacklist(HTMLFILE)
+                    if EnableHTML == "yes":
+                        if OldRoot in os.path.join(Root , file):
+                            pass
+                        else:
+                            _ , GameName , Maybethis = Root.rsplit("/",2)
+                            OldRoot = Maybethis
+                            HTMLFILE = (os.path.join(Root , file))
+                            InBlacklist(HTMLFILE)
                 if Proton == "yes":
                     if file.endswith(".exe"):
                         ExeFile = (os.path.join(Root , file))
@@ -369,6 +380,10 @@ def getfiles(scandir):
                 else:
                     pass
             if OS == "Darwin":# in macOS .app files are treated as folders for NO REASON!
+                if file == 'index.html':
+                    if EnableHTML == "yes":
+                        HTMLFILE = (os.path.join(Root , file))
+                        InBlacklist(HTMLFILE)
                 if ".app" in Root:
                     NoRepeasts = 1
                     for i in SkipPath.split("\n"):
@@ -427,6 +442,51 @@ def AddShortcut(File):
         if OS == "Darwin":
             Run = ('"%s/shortcuts.vdf" %s "" %s 0 1 1 %s 0 "Non-Steam-Game"'%(ReplaceVDF , split_path(File) , LaunchOptions , inVRLibrary))
             os.system("python3 shortcuts.py %s"%Run)
+def AddHTMLGame(gamedir):
+    global emptyport
+    if OS == 'Windows':
+        emptyport += 1
+        HTMLServerLaunch1 = "'C:\\Windows\\System32\\cmd.exe' /c start /b "
+        HTMLServerLaunch2 = 'python -m http.server %d -d '%emptyport
+        HTMLGameLaunch = ' & start /max http://127.0.0.1:%d'%emptyport
+        name , _ = gamedir.split(".")
+        S , name , _ = name.rsplit("\\",2)
+        start = os.path.join(S , name)
+        if " " in start:
+            FullHTML = ("%s%s'%s'%s"%(HTMLServerLaunch1 , HTMLServerLaunch2 , start , HTMLGameLaunch))
+        else:
+            FullHTML = ("%s%s%s%s"%(HTMLServerLaunch1 , HTMLServerLaunch2 , start , HTMLGameLaunch))
+        return ('"%s" "%s" "%s" "%s"'%(name , FullHTML , start , gamedir))
+    if OS == 'Linux':
+        emptyport += 1
+        HTMLServerLaunch = 'python3 -m http.server %d -d '%emptyport
+        HTMLGameLaunch = ' & xdg-open http://0.0.0.0:%d'%emptyport
+        _ , name , _ = gamedir.rsplit("/",2)
+        start , _= gamedir.rsplit("/", 1)
+        if " " in start:
+            FullHTML = ("%s'%s'%s"%(HTMLServerLaunch , start , HTMLGameLaunch))
+        else:
+            FullHTML = ("%s%s%s"%(HTMLServerLaunch , start , HTMLGameLaunch))
+        return ('"%s" "%s" "%s" "%s"'%(name , FullHTML , start , gamedir))
+    if OS == 'Darwin':
+        def MakeSHscript(name , FullHTML , start , gamedir):
+            shfile = '%s/%s.sh'%(start,name)
+            WriteSH = open(shfile,'w')
+            SHContents = FullHTML
+            WriteSH.write('#!/bin/sh\n%s'%SHContents)
+            st = os.stat(shfile)
+            os.chmod(shfile,st.st_mode|stat.S_IEXEC)
+            return ('"%s" "%s" "%s" "%s"'%(name , shfile , start , gamedir))
+        emptyport += 1
+        HTMLServerLaunch = "python3 -m http.server %d -d "%emptyport
+        HTMLGameLaunch = ' & open http://127.0.0.1:%d\n$SHELL'%emptyport
+        _ , name , _ = gamedir.rsplit("/",2)
+        start , _= gamedir.rsplit("/", 1)
+        if " " in start:
+            FullHTML = ("%s'%s'%s"%(HTMLServerLaunch , start , HTMLGameLaunch))
+        else:
+            FullHTML = ("%s%s%s"%(HTMLServerLaunch , start , HTMLGameLaunch))
+        return(MakeSHscript(name , FullHTML , start , gamedir))
 def ClearCLI():
     if OS == "Linux" or "Darwin":
         os.system('clear')
@@ -434,9 +494,9 @@ def ClearCLI():
         os.system('cls')
 def GUI():
     getsettings()
-    def WriteSettings(SteamID , InstallLocation , DefaultCleanout , Proton):
+    def WriteSettings(SteamID , InstallLocation , DefaultCleanout , Proton , EnableHTML):
         SettingsFile = open('info/settings','w')
-        SettingsFile.write('%s , %s , %s , %s'%(SteamID , InstallLocation , DefaultCleanout , Proton))
+        SettingsFile.write('%s , %s , %s , %s , %s'%(SteamID , InstallLocation , DefaultCleanout , Proton , EnableHTML))
         return
     def getDir(NumedList , DirToRM , AddNum):
         _ , RMDIR = NumedList.split('%s) '%DirToRM)
@@ -523,10 +583,10 @@ def GUI():
     Layer1 = 1
     while Layer1 == 1:
         ClearCLI()
-        WhichSetting = input('What would you like to do?\n\n(1)Enable/disable shortcut cleaning\n\n(2)Enable/Disable Proton shortcuts(.exe games on linux)\n\n(3)Manage folders to scan\n\n(4)Add shortcuts(will quit after done adding)\n\n(5)exit\n')
+        WhichSetting = input('What would you like to do?\n\n(1)Enable/disable shortcut cleaning\n\n(2)Enable/Disable Proton shortcuts(.exe games on linux)\n\n(3)Enable/Disable HTML games(almost done but not exstensively bug tested)\n\n(4)Manage folders to scan\n\n(5)Add shortcuts(will quit after done adding)\n\n(6)exit\n')
         ReadSettings = open('info/settings','r')
         ReadSettings.seek(0)
-        SteamID , InstallLocation , DefaultCleanout , Proton = ReadSettings.read().split(' , ')
+        SteamID , InstallLocation , DefaultCleanout , Proton , EnableHTML = ReadSettings.read().split(' , ')
         ValidOptions = ('1','2','3','4','5')
         for i in ValidOptions:
             if WhichSetting == i:
@@ -549,7 +609,7 @@ def GUI():
                             if ChangeCleanout == 'n':
                                 Layer2 = 0
                                 pass
-                        WriteSettings(SteamID , InstallLocation , DefaultCleanout , Proton)
+                        WriteSettings(SteamID , InstallLocation , DefaultCleanout , Proton , EnableHTML)
                     if WhichSetting == '2':
                         if OS != 'Linux':
                             input('This is a linux option only(press enter to continue)')
@@ -569,16 +629,33 @@ def GUI():
                                     Layer2 = 0
                                 if ChangeProton == 'n':
                                     Layer2 = 0
-                            WriteSettings(SteamID , InstallLocation , DefaultCleanout , Proton)
+                            WriteSettings(SteamID , InstallLocation , DefaultCleanout , Proton , EnableHTML)
                     if WhichSetting == '3':
+                        if EnableHTML == 'yes':
+                            ChangeHTML = input('HTML5 games will be added to steam\n\nWould you like to change that?(y/n)')
+                        elif EnableHTML == 'no':
+                            ChangeHTML = input('HTML5 games will not be added to steam\n\nWould you like to change that?(y/n)')
+                        if ChangeHTML == 'y' or 'n':
+                            ClearCLI()
+                            if ChangeHTML == 'y':
+                                if EnableHTML == 'yes':
+                                    EnableHTML = 'no'
+                                elif EnableHTML == 'no':
+                                    EnableHTML = 'yes'
+                                Layer2 = 0
+                            if ChangeHTML == 'n':
+                                Layer2 = 0
+                                pass
+                        WriteSettings(SteamID , InstallLocation , DefaultCleanout , Proton , EnableHTML)
+                    if WhichSetting == '4':
                         DirManager()
                         Layer2 = 0
-                    if WhichSetting == '4':
+                    if WhichSetting == '5':
                         Layer2 = 0
                         Layer1 = 0
                         ClearCLI()
                         run()
-                    if WhichSetting == '5':
+                    if WhichSetting == '6':
                         Layer1 = 0
                         Layer2 = 0         
 def run():
